Refactor test files structure to be consistent across projects (#37)
