+++ dkg/src/tests.rs
@@ -1,2 +1,1 @@
-mod inputs_tests;
-mod integration_test;
+mod inputs;
