+++ dkg/tests.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod integration_tests;
