+++ trusted-dealer/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod cli;
+pub mod inputs;
+pub mod trusted_dealer_keygen;
