+++ trusted-dealer/src/main.rs
@@ -1,6 +1,5 @@
 mod cli;
 mod inputs;
-mod output;
 mod trusted_dealer_keygen;
 
 #[cfg(test)]
