+++ trusted-dealer/src/tests.rs
@@ -1,3 +1,2 @@
-mod inputs_tests;
-mod integration_test;
-mod output_tests;
+mod inputs;
+mod print_values;
