+++ trusted-dealer/src/tests/inputs.rs
