+++ trusted-dealer/tests/helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod signature_gen;
