+++ trusted-dealer/tests/helpers/signature_gen.rs
